Move all form pages over to using FormPages module instead of individual CSS pages
