+++ frontend/src/css/FormPages.module.css
@@ -8,13 +8,7 @@
     color: #fdfdfd;
 }
 
-.jobSubmissionHeader {
-    color: #002337;
-    margin: 2rem;
-    text-align: center;
-}
-
-.jobAppForm {
+.form {
     background-color: #002337;
     display: flex;
     flex-direction: column;
@@ -27,11 +21,6 @@
     display: flex;
     gap: 2rem;
     justify-content: space-between;
-}
-
-input, label, select, option {
-    font-family: 'Lucida Sans';
-    font-size: 1.1rem;
 }
 
 .submitForm {
@@ -50,4 +39,9 @@ input, label, select, option {
 
 .formNote {
     font-family: 'Lucida Sans';
+}
+
+input, label, select, option {
+    font-family: 'Lucida Sans';
+    font-size: 1.1rem;
 }
