--- frontend/src/css/Login.module.css
@@ -1,48 +1,0 @@
-.header {
-    color: #002337;
-    margin: 2rem;
-    text-align: center;
-}
-
-.loginFormSection {
-    min-height: 26rem;
-    background-color: #0077b6;
-    padding: 4rem;
-    margin-bottom: 3.5rem;
-    display: flex;
-    justify-content: center;
-    color: #fdfdfd;
-}
-
-.loginForm {
-    background-color: #002337;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: 1rem;
-    padding: 5rem;
-}
-
-.formOption {
-    display: flex;
-    gap: 2rem;
-    justify-content: space-between;
-}
-
-.submitForm {
-    border: none;
-    width: 10rem;
-    padding: 1rem;
-    background-color: #0077b6;
-    color: #fdfdfd;
-    font-size: 1rem;
-    cursor: pointer;
-    border-radius: 10px;
-    margin: auto;
-    margin-top: 2rem;
-    font-family: 'Lucida Sans';
-}
-
-.formNote {
-    font-family: 'Lucida Sans';
-}
